requirements.txt
torch
torchvision
torchaudio
pillow
tqdm
pandas
scikit-learn
matplotlib
seaborn
requests